Major Refactor WIP Goals: - Code cleanup to better reflect current code quality - Mirror the older version's UI and functionality as much as possible - Refactor app project to utilize Kotlin, Coroutines, Jetpack Compose, Hilt/Dagger, etc.
Updates:
- Now utilizing fragments in audio selection [main] activity for audio selection and media playback controls functionality
- Renamed MusicPlayer___.kt files to MediaPlayer___.kt
- Removed more old/unnecessary code
- Imported ConstraintLayout

Notes/TODO:
- ✅ Core media player logic functionality has been implemented
- ✅ Core audio selection functionality has been implemented
- ✅ Core permission-check functionality has been implemented
- ✅ Core limited media retrieval functionality has been implemented
- Need to implement toolbar/top bar
- Need to implement navigation to Now Playing screen
- ✅ Need to utilize fragments in single activity for audio selection functionality
- [Nice to have] Need to implement loop functionality
- [Nice to have] Need to implement audio list items filtering
- Consider segregating MediaPlayer implementation, state and functionality out MusicPlayerStateViewModel and into a "MediaPlayerService"
- ✅ Rename MusicPlayer___.kt files to MediaPlayer___.kt
- Finish cleaning up/removing old code

Known Bugs:
- ContentResolver appears to only be retrieving audio files from internal storage, not SDCard - even when referencing external URI path (uncommenting the internal URI portion will retrieve notification and ringtone files)
- [Warning - Watch-Item] MusicPlayerStateViewModel MediaPlayer state handling is 'delicate'. I have seen ANRs due to mishandling (probably due to the 'nowPlayingTrackProgress' while-loop?). Keeping this note as a watch-item.
- Slider's progress bar (thumb) not tracking user's continuous input - instead, it only updates after motion release
