+++ app/build.gradle
@@ -83,6 +83,7 @@ dependencies {
     implementation composeBom
     androidTestImplementation composeBom
 
+    implementation 'androidx.constraintlayout:constraintlayout:2.2.0-alpha13'
     implementation 'androidx.compose.material3:material3'
     // Android Studio Preview support
     implementation 'androidx.compose.ui:ui-tooling-preview'
