+++ app/src/main/AndroidManifest.xml
@@ -26,15 +26,6 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.activities.MainActivity"
-            android:exported="true"
-            android:label="PROJECTMusicApp">
-<!--            <intent-filter>-->
-<!--                <action android:name="android.intent.action.MAIN" />-->
-<!--                <category android:name="android.intent.category.LAUNCHER" />-->
-<!--            </intent-filter>-->
-        </activity>
-        <activity
             android:name=".ui.activities.SelectedSongActivity"
             android:label="Details"
             android:screenOrientation="fullSensor"
