--- app/src/main/res/layout/custom_list_item.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/main_parent"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:minHeight="@dimen/list_item_height"
-    android:orientation="horizontal"
-    android:background="@drawable/button_bg">
-    <ImageView
-        android:id="@+id/album_artwork"
-        android:layout_width="@dimen/list_item_height"
-        android:layout_height="@dimen/list_item_height"
-        android:minHeight="@dimen/list_item_height"
-        android:layout_margin="4dp"
-        android:src="@drawable/default_icon_song_album1" />
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-        <TextView
-            android:id="@+id/song_title"
-            style="@style/custom_item_textview"
-            android:textSize="16sp"
-            android:gravity="center_vertical"
-            android:textStyle="bold"
-            android:textColor="@android:color/white"/>
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-            <TextView
-                android:id="@+id/artist_and_album_info"
-                style="@style/custom_item_textview"
-                android:textSize="12sp"
-                android:textColor="@android:color/white"/>
-        </LinearLayout>
-    </LinearLayout>
-</LinearLayout>
